+++ db/migrate/20230121013252_rename_amenity_to_amenity_pass.rb
@@ -1,0 +1,5 @@
+class RenameAmenityToAmenityPass < ActiveRecord::Migration[7.0]
+  def change
+    rename_table :amenities, :amenity_passes
+  end
+end